refactor(logging): extract logging init code into shared module for patch evaluation component and llm dispatch component
