+++ src/autopatchshared/autopatchshared/__init__.py
@@ -1,0 +1,1 @@
+from .init_logging import init_logging
