+++ src/autopatchshared/autopatchshared/init_logging.py
@@ -1,0 +1,53 @@
+import os
+import json
+import logging
+import logging.config
+
+
+def init_logging(logging_config: str, appname: str) -> logging.Logger:
+    """
+    Initializes logging from a JSON configuration file.
+
+    If the JSON file cannot be loaded or the configuration is invalid,
+    a basic logging configuration is used as a fallback.
+
+    Parameters:
+        logging_config (str): Path to the JSON file with logging configuration.
+        appname (str): Name of the application logger.
+
+    Returns:
+        logging.Logger: Configured logger instance.
+    """
+    try:
+        # Check if the configuration file exists
+        if not os.path.exists(logging_config):
+            raise FileNotFoundError(
+                f"Logging configuration file '{logging_config}' not found."
+            )
+
+        # Load the JSON configuration from the file
+        with open(logging_config, "r") as config_file:
+            config_dict = json.load(config_file)
+
+        # Apply the logging configuration
+        logging.config.dictConfig(config_dict)
+
+    except FileNotFoundError as fnf_error:
+        print(fnf_error)
+        print("Falling back to basic logging configuration.")
+        logging.basicConfig(level=logging.INFO)
+
+    except json.JSONDecodeError as json_error:
+        print(f"Error decoding JSON from {logging_config}: {json_error}")
+        print("Falling back to basic logging configuration.")
+        logging.basicConfig(level=logging.INFO)
+
+    except Exception as e:
+        print(f"Unexpected error while loading logging configuration: {e}")
+        print("Falling back to basic logging configuration.")
+        logging.basicConfig(level=logging.INFO)
+
+    # Get and return the logger for the specified application name
+    logger = logging.getLogger(appname)
+    logger.info("Logger initialized successfully.")
+    return logger
