+++ src/autopatchshared/autopatchshared/observability.py
@@ -1,0 +1,3 @@
+def init_open_telemetry():
+    print("Not implemented yet")
+    pass
