+++ src/fuzzing-service/Dockerfile
@@ -20,6 +20,7 @@ RUN groupadd -r appuser && \
 
 COPY requirements.txt .
 COPY autopatchdatatypes-0.2.3-py3-none-any.whl .
+COPY autopatchshared-0.3.0-py3-none-any.whl .
 RUN pip install --break-system-packages --user --no-cache-dir -r requirements.txt
 
 COPY fuzzing_service.py /app/fuzzing_service.py
