+++ src/fuzzing-service/fuzzing_service.py
@@ -16,6 +16,7 @@ from typing import Final, List
 import paho.mqtt.client as mqtt_client
 import paho.mqtt.enums as mqtt_enums
 from autopatchdatatypes import CrashDetail
+from autopatchshared import init_logging
 from cloudevents.http import CloudEvent
 
 # this is the name of the environment variable that will be used point to the configuration map file to load
@@ -52,55 +53,6 @@ class Config:
     message_broker_ca_certs: str
     message_broker_certfile: str
     message_broker_keyfile: str
-
-
-def init_logging(logging_config: str, appname: str) -> logging.Logger:
-    """
-    Initializes logging from a JSON configuration file.
-
-    If the JSON file cannot be loaded or the configuration is invalid,
-    a basic logging configuration is used as a fallback.
-
-    Parameters:
-        logging_config (str): Path to the JSON file with logging configuration.
-        appname (str): Name of the application logger.
-
-    Returns:
-        logging.Logger: Configured logger instance.
-    """
-    try:
-        # Check if the configuration file exists
-        if not os.path.exists(logging_config):
-            raise FileNotFoundError(
-                f"Logging configuration file '{logging_config}' not found."
-            )
-
-        # Load the JSON configuration from the file
-        with open(logging_config, "r") as config_file:
-            config_dict = json.load(config_file)
-
-        # Apply the logging configuration
-        logging.config.dictConfig(config_dict)
-
-    except FileNotFoundError as fnf_error:
-        print(fnf_error)
-        print("Falling back to basic logging configuration.")
-        logging.basicConfig(level=logging.INFO)
-
-    except json.JSONDecodeError as json_error:
-        print(f"Error decoding JSON from {logging_config}: {json_error}")
-        print("Falling back to basic logging configuration.")
-        logging.basicConfig(level=logging.INFO)
-
-    except Exception as e:
-        print(f"Unexpected error while loading logging configuration: {e}")
-        print("Falling back to basic logging configuration.")
-        logging.basicConfig(level=logging.INFO)
-
-    # Get and return the logger for the specified application name
-    logger = logging.getLogger(appname)
-    logger.info("Logger initialized successfully.")
-    return logger
 
 
 def load_config() -> dict:
