+++ src/fuzzing-service/requirements.txt
@@ -1,4 +1,4 @@
 paho-mqtt==2.1.0
-# autopatchdatatypes==0.2.3
-autopatchdatatypes-0.2.3-py3-none-any.whl
+autopatchdatatypes-0.3.0-py3-none-any.whl
+autopatchshared-0.3.0-py3-none-any.whl
 cloudevents==1.11.0
